Move getAttributableUrl to utils/task.js
